+++ src/main/java/hu/lae/riskparameters/RiskParameters.java
@@ -1,7 +1,5 @@
 package hu.lae.riskparameters;
 
-import java.math.BigDecimal;
-
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
@@ -11,7 +9,7 @@ public class RiskParameters {
     
     public final String name;
     
-    public final BigDecimal amortizationRate;
+    public final double amortizationRate;
     
     public final Haircuts haircuts;
     
@@ -19,10 +17,10 @@ public class RiskParameters {
     
     public final InterestRate longTermInterestRate;
     
-    public final BigDecimal dscrThreshold;
+    public final double dscrThreshold;
 
-    public RiskParameters(String id, String name, BigDecimal amortizationRate, Haircuts haircuts, InterestRate shortTermInterestRate,
-            InterestRate longTermInterestRate, BigDecimal dscrThreshold) {
+    public RiskParameters(String id, String name, double amortizationRate, Haircuts haircuts, InterestRate shortTermInterestRate,
+            InterestRate longTermInterestRate, double dscrThreshold) {
         this.id = id;
         this.name = name;
         this.amortizationRate = amortizationRate;
